Docs Refactor
This patch is documentation only.  The changes are not complete and information
in the files should be updated as the progress progresses.  The goal is NOT
to update the files, but to move them into better locations as dicussed in
weekly calls.

The change includes making the root readme more general with links to deeper
documents.  Landing pages for running RefStack and TCUP have been moved to the
/doc directory.

Change-Id: Ifdf4afe0411f92253e0b4cb3ad64a0e6be3bd172