Add ability to build basic configuration from environment variables if they are passed.
For now, just using the most basic variables.  This could be expanded
later but this should be sufficient for initial testing.

The expected env variables are included in the attached doc.

This configuration exposes assumptions in the rest of the file
where more config data is expected than should be required.  The goal
of this patch is to ADD ENV CONFIG not to correct how the config
data is consumed in the file.

Change-Id: I849ec113c00b46c7bb40d748e1d88a5d70f03dc7