Temporary fix for get-script to use the new execute_test.
In https://review.openstack.org/#/c/86419/,
tools/execute_test.py has been moved to tools/execute_test/execute_test.
This change fixes the get-script API to return the execute_test at the new location.
This is a temporary fix so that people can continue to collect data using the
local Refstack installation with Docker.
For the future, there is a proposal to remove the get-script API and replaced it
with getting the execute_test from github directly.

Change-Id: I6aff5a1f00dbc18b6389a04210f883a7e3520769