Controlled JSON files for Defcore management
The driver file in this pull are placeholders
for the actual content in the files.

In the case of the coretests and capabilities list,
preliminary values have been provided for discussion
and review of the schema.

** THEY ARE NOT OFFICIAL RESULTS **

These files will contain approved information that
is managed by the DefCore committee relating to
core and driver test.

For more information about this process please
review https://wiki.openstack.org/wiki/Governance/CoreCriteria
and Rob Hirschfeld's post: http://robhirschfeld.com/tag/defcore/

Changes to these files must be +1 by a DefCore chair.

Change-Id: I7ebc6f482f062725ce2cdb83a5102eaba05d8cfe