Report Visualization Spec and Mocks
For https://storyboard.openstack.org/#!/story/111

This spec should be considered complete for final review.

This is a pair of read only screens for helping
community compare tests results starting from the
DefCore capabilities and allowing them to inspect
specific items on drill down.

The focus is on comparison against core and
other clouds not on quantative analysis.

Change-Id: I85fa223a4fd67b16d9103593f727a6811c741452