Define Reports for DefCore.
For story https://storyboard.openstack.org/#!/story/109

This spec should be considered complete for final review.

It is a basic read only screen for rendering test results
in a way that helps the DefCore committee (and other
community members) find tests & capabilities that should
be added to core or should be removed from core.

Change-Id: I879b73cba32d8ae65aa50ad5cb42cc7c9baa5647