Use fully qualified test id in the coretests.json file.
The coretests.json file includes a list of tests that are defined as "core" by
the DefCore committee.  Currently, the coretests.json file (in the
defcore/havana directory) uses the method names defined in the Tempest Python
test classes as the test names.  While these method names are unique in Havana,
it is not the case in Icehouse where some of the method names are being used by
multiple test classes of different OpenStack components.

Change-Id: Idcc62c43984356fca00464dd192ae1292569e0b8