Complete Havana capabilities as passed by the Board in July 2014
This change includes capabilities descriptions from Troy,
  Sergey, Julia & Rocky.

PLEASE DO NOT ASK FOR DESCRIPTION OR NAME CHANGES...
+1 the file and then submit patches against it
we know that the tests, names and descriptions are incomplete.

The tests are NOT included in the capabilities because more
review is needed.  We expect that they will be updated
after the schema with capabilities are approved.

Change-Id: I81016dbb25bf421723e89f2e2a2b21466d461605