Add CORS hook to api application
A new config option was added called 'allowed_cors_origins' which
takes a list of domains where Cross Origin Resource Sharing is
permitted. If a request is in this list of sites, CORS headers
are added  to the responses by the hook in order to allow resource sharing.
This is needed by the UI to be able to retrieve data from the API.

Change-Id: I7d98d15b05e8f0da8af3a24a424779aa5be788d5