Add results and test report page
This is the groundwork for the community results listing page
and the test run report page. The report page is fairly basic, primarily
showing how results stack up against defcore capabilities. Design is subject
to change, but this gets the ball rolling. A config.json.sample was also added.
The UI now expects a config.json in the root which will contain the
Refstack API URL.

Change-Id: Id7a376d0bccda5cbb5daf05e52a2c174ad40b497