Added PhantomJS test execution
This patch adds test execution in the headless phantomjs runner, a
feature requested by package maintainers during the summit (for all
javascript projects, not refstack explicitly). The phantomjs
requirement has been explicitly added, as at this point there are
several upstream changes pending inside of npm that would make
this build brittle.

1- NPMv3 will no longer automatically install peerDependencies.
2- karma-phantomjs-launcher will soon gracefully degrade to a
system-installed phantomjs if it is not otherwise available.

Once both of these land we can remove the explicit reference to
PhantomJS.

Change-Id: I054792c23e1f38538800c901605a032118184925