Check for RSA key existence before accepting uploading data.
RefStack recently added features to enable the uploading of data with key.
Currently, RefStack accepts the uploaded data regardless of whether
the keys exist in RefStack or not. This spec describes the validation
process update needed to ensure that RefStack only accepts data with
those keys that are previously imported into RefStack.

Change-Id: I5bb7de969fe92258d54a64feec01310afcf91038