Set pool_recycle for beaker sessions
Occasionally a beaker database connection will be disconnected
due to a long period of database inactivity. This can cause
users to be logged out before a session expires and can also
cause internal server errors when users try to log in. This patch
adds a sqlalchemy pool_recycle value of 10 minutes, so that beaker
doesn't potentially try to use a connection that has been disconnected
by the mysql server.

Change-Id: Ida99efbba1e99e72c1610cd2335560706f1523d7
Closes-Bug: #1514290