add 'is_admin' flag to profile result
rendering of some pages (vendors, clouds, software, ...)
needs to know which UI elements should be present on the page.
instead of returning this flag in many requests I suggest to
make it available in 'auth.current_user' structure on client side.
This flag means that current user is in Foundation group.
It will influence only to pages rendering.

Change-Id: If8e993495bb301a4272532b6ecce6e1d4fdaa21e