Add option to view/download test lists
On the guideline page when viewing DefCore capabilities, users
might want to get the test list containing only required or advisory
tests. This patch provides a way for a user to get that test list
to use with RefStack client. The phantomjs version was updated to allow
testing with Blobs.

Change-Id: I9243f5781cb8a688ca3e64645302db5bc4527c37