Make profile endpoint not error out
This patch makes it so that the check for if a user is a foundation
admin will not error out when a foundation org does not exist. Instead
an empty list is returned, saying that there are no foundation users.

The documentation was also updated to give instructions for creating the
Foundation group/organization, and adding a user to it.

Closes-Bug: #1552548

Change-Id: If675d856b69ab854a3daf296f81050525ba5d50b