Refactor capabilities
Rename the 'capabilities' endpoint to 'guidelines' to align more
with DefCore terminology. Several references to 'capabilities' were
thus updated to 'guidelines'.
Split out the actual requests to the DefCore repository so
that it can be used by other components.
Also, the cache expiration time was increased to 12 hours from 10 minutes.

Change-Id: I031798ffc32cdd66428af83a1ee610f00385c524