Add next.json to DefCore guideline list
next.json is the proposed guideline for DefCore, and is published
by the working group for comment and feedback. As part of the
feedback process, users and DefCore working group members need
to evaluate the proposed standard against existing test results.

This change modifies the guideline pattern matcher to pull in
the next file as well as the approved guidelines.

Change-Id: I8c8391b9ab14d1c6d7286710b5e64b72be3e92a8