Add verification_status field to test
This patch adds the ability for Foundation admins
to mark a test that has been associated to a target program and
guideline as 'verified' so that the test can not be deleted or updated.

Updated error messages to reflect that products/vendors can not
be deleted if a test is associated.

Change-Id: Ifa7cc2eff9d2bce9129ec1738adebfa3a0a4966f
Implements-Spec: https://review.openstack.org/#/c/343954