Disallow deleting the null version of a product
The null version is implicitly created with the product and should
only be deleted if the product is deleted. This version is used
when a user doesn't want to associate a test to an explicit version
but just the product itself. This patch restricts users from
explicitly deleting this null version.

Change-Id: Ibb29940b5702f73cd50ed2a29e5a23a4d87ed887