Rename key name from 'reason' to 'registration_decline_reason'
Currently whenever Foundation admins decline a vendor registration
request.  A key/value pair will be added to the vendor properties field
with the key name as 'reason'.  Since the key/value data will be shown
as-is in the response from the 'v1/vendors/{vendor_id} API call. This
patch renames the key name to a more descriptive one.

Change-Id: I6a43f7671a61a05bb2f0186dc52031eb6760eb59