Create spec for optional subunit data import
Re: primary assignee: I know Chris had mentioned he wanted me
working on this; primary assignee can be changed if somebody
else is going to be the main actor in this change.

This is very much a work in progress. Some topics currently
being researched/up for debate:
* Is subunit2sql's input sanitization adequate for our needs,
  where protection against malicious file upload is concerned?
* Is there a way to store the data in subunit2sql's db schema
  in the RefStack db or vice versa?

Sections that I will be adding more detail to shortly:
* REST API impact
* Other deployer impact

Change-Id: I03e212694d7d7a586f1caabedbebd7da6a0f38b9