Set a custom alembic_version for refstack
This is the update to the patch by mtreinish that allows for the
setting of a custom alembic version_table name. In this patch,
the name of this table is set via a new flag in conf. This is the
first patch of several that will allow us to build in the ability
to collect and upload subunit data to the refstack server. By
default, the table is set to be named "alembic_version".

Previously, this patch also included some alembic configuration.
This has been removed due to the fact that renaming the
alembic_version table is not as simple as previously hoped, and
will, as a result, require the usage of a migration script.
Because of this, I have chosen to sever the two patches, largely,
due to the fact that they have separate functionalities, although
the migration patch will consume the conf setting created in this
patch, so this merging this patch first would be highly advisable.
Change-Id: I4c82aabb727694d7cd55aaab339bbc55b1835f27