Update deps so that refstack builds and deploys
In bindep we specify we need GCC to compile deps and set tags to compile
and test as appropriate as none of these packages are needed as runtime
dependencies. In requirements.txt we unpin alembic and beaker and add
PyMySQL. Alembic and beaker need to be unpinned to work with modern
python3 and PyMySQL is a python only runtime dep if using MySQL.

This makes building docker images much simpler.

Change-Id: I18a296de190324c545feda3b1b87e0a9424d0f96